Junit assertions

+++ PHASE4/junit-demo/src/main/java/com/sl/Calculator.java
@@ -9,7 +9,7 @@ public class Calculator {
 	
 	
 	public float divide(int a, int b){		
-		return (float) a/ (float) b;		
+		return  a/ b;		
 	}
